Add some structured error messages, organization
