+++ saml.go
@@ -1,18 +1,6 @@
 package saml2
 
-import (
-	"bytes"
-	"compress/flate"
-	"encoding/base64"
-	"errors"
-	"fmt"
-	"net/url"
-	"time"
-
-	"github.com/beevik/etree"
-	"github.com/russellhaering/goxmldsig"
-	"github.com/satori/go.uuid"
-)
+import dsig "github.com/russellhaering/goxmldsig"
 
 const issueInstantFormat = "2006-01-02T15:04:05"
 
@@ -30,172 +18,6 @@ type SAMLServiceProvider struct {
 
 func (sp *SAMLServiceProvider) SigningContext() *dsig.SigningContext {
 	return dsig.NewDefaultSigningContext(sp.SPKeyStore)
-}
-
-func (sp *SAMLServiceProvider) BuildAuthRequest() (string, error) {
-	authnRequest := &etree.Element{
-		Space: "samlp",
-		Tag:   "AuthnRequest",
-	}
-
-	authnRequest.CreateAttr("xmlns:samlp", "urn:oasis:names:tc:SAML:2.0:protocol")
-	authnRequest.CreateAttr("xmlns:saml", "urn:oasis:names:tc:SAML:2.0:assertion")
-
-	authnRequest.CreateAttr("ID", "_"+uuid.NewV4().String())
-	authnRequest.CreateAttr("Version", "2.0")
-	authnRequest.CreateAttr("ProtocolBinding", "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST")
-	authnRequest.CreateAttr("AssertionConsumerServiceURL", sp.AssertionConsumerServiceURL)
-	authnRequest.CreateAttr("AssertionConsumerServiceIndex", "0")
-	authnRequest.CreateAttr("AttributeConsumingServiceIndex", "0")
-	authnRequest.CreateAttr("IssueInstant", time.Now().UTC().Format(issueInstantFormat))
-
-	authnRequest.CreateAttr("Destination", "http://idp.astuart.co/idp/profile/SAML2/Redirect/SSO")
-	// authnRequest.CreateAttr("Destination", sp.IdentityProviderSSOURL)
-
-	authnRequest.CreateElement("saml:Issuer").SetText(sp.IdentityProviderIssuer)
-
-	nameIdPolicy := authnRequest.CreateElement("samlp:NameIDPolicy")
-	nameIdPolicy.CreateAttr("AllowCreate", "true")
-	nameIdPolicy.CreateAttr("Format", sp.NameIdFormat)
-
-	requestedAuthnContext := authnRequest.CreateElement("samlp:RequestedAuthnContext")
-	requestedAuthnContext.CreateAttr("Comparison", "exact")
-
-	authnContextClassRef := requestedAuthnContext.CreateElement("saml:AuthnContextClassRef")
-	authnContextClassRef.SetText("urn:oasis:names:tc:SAML:2.0:ac:classes:PasswordProtectedTransport")
-
-	doc := etree.NewDocument()
-
-	if sp.SignAuthnRequests {
-		signed, err := sp.SigningContext().SignEnveloped(authnRequest)
-		if err != nil {
-			return "", err
-		}
-
-		doc.SetRoot(signed)
-	} else {
-		doc.SetRoot(authnRequest)
-	}
-
-	buf := &bytes.Buffer{}
-
-	_, err := doc.WriteTo(buf)
-	if err != nil {
-		return "", err
-	}
-
-	return doc.WriteToString()
-}
-
-func (sp *SAMLServiceProvider) BuildAuthURL(relayState string) (string, error) {
-	parsedUrl, err := url.Parse(sp.IdentityProviderSSOURL)
-	if err != nil {
-		return "", err
-	}
-
-	authnRequest, err := sp.BuildAuthRequest()
-	if err != nil {
-		return "", err
-	}
-
-	buf := &bytes.Buffer{}
-
-	fw, err := flate.NewWriter(buf, flate.DefaultCompression)
-	if err != nil {
-		return "", fmt.Errorf("flate NewWriter error: %v", err)
-	}
-
-	_, err = fw.Write([]byte(authnRequest))
-	if err != nil {
-		return "", fmt.Errorf("flate.Writer Write error: %v", err)
-	}
-
-	err = fw.Close()
-	if err != nil {
-		return "", fmt.Errorf("flate.Writer Close error: %v", err)
-	}
-
-	qs := parsedUrl.Query()
-
-	qs.Add("SAMLRequest", base64.StdEncoding.EncodeToString(buf.Bytes()))
-
-	if relayState != "" {
-		qs.Add("RelayState", relayState)
-	}
-
-	parsedUrl.RawQuery = qs.Encode()
-	return parsedUrl.String(), nil
-}
-
-func (sp *SAMLServiceProvider) Validate(el *etree.Element) error {
-	el = el.Copy()
-
-	destinationAttr := el.SelectAttr(DestinationAttr)
-	if destinationAttr == nil {
-		return errors.New("Missing Destination attribute")
-	}
-	if destinationAttr.Value != sp.AssertionConsumerServiceURL {
-		return errors.New(fmt.Sprintf("Did not recognize Destination value, Expected: %s, Actual: %s", sp.AssertionConsumerServiceURL, destinationAttr.Value))
-	}
-
-	idAttr := el.SelectAttr(IdAttr)
-	if idAttr == nil || idAttr.Value == "" {
-		return errors.New("Missing ID attribute")
-	}
-
-	versionAttr := el.SelectAttr(VersionAttr)
-	if versionAttr == nil {
-		return errors.New("Missing Version attribute")
-	}
-	if versionAttr.Value != "2.0" {
-		return errors.New("Unsupported SAML version")
-	}
-
-	assertionElement := el.FindElement(AssertionTag)
-	if assertionElement == nil {
-		return ErrMissingAssertion
-	}
-
-	subjectStatement := assertionElement.FindElement(childPath(assertionElement.Space, SubjectTag))
-	if subjectStatement == nil {
-		return errors.New("Missing Subject")
-	}
-
-	subjectConfirmationStatement := subjectStatement.FindElement(childPath(assertionElement.Space, SubjectConfirmationTag))
-	if subjectConfirmationStatement == nil {
-		return errors.New("Missing SubjectConfirmation")
-	}
-
-	methodAttr := subjectConfirmationStatement.SelectAttr(MethodAttr)
-	if methodAttr.Value != "urn:oasis:names:tc:SAML:2.0:cm:bearer" {
-		return errors.New("Unsupported subject confirmation method")
-	}
-
-	subjectConfirmationDataStatement := subjectConfirmationStatement.FindElement(childPath(assertionElement.Space, SubjectConfirmationDataTag))
-	if subjectConfirmationDataStatement == nil {
-		return errors.New("Missing SubjectConfirmationData")
-	}
-
-	recipientAttr := subjectConfirmationDataStatement.SelectAttr(RecipientAttr)
-	if recipientAttr == nil {
-		return errors.New("Missing Recipient attribute")
-	}
-	if recipientAttr.Value != sp.AssertionConsumerServiceURL {
-		return errors.New(fmt.Sprintf("Did not recognize Recipient value, Expected: %s, Actual: %s", sp.AssertionConsumerServiceURL, recipientAttr.Value))
-	}
-
-	notOnOrAfterAttr := subjectConfirmationDataStatement.SelectAttr(NotOnOrAfterAttr)
-	if notOnOrAfterAttr != nil {
-		after, err := time.Parse(time.RFC3339, notOnOrAfterAttr.Value)
-		if err != nil {
-			return errors.New("Could not parse 'NotOnOrAfter' attribute")
-		}
-		if time.Now().After(after) {
-			return errors.New("SubjectConfirmationData is no longer valid")
-		}
-	}
-
-	return nil
 }
 
 type ProxyRestriction struct {
